+++ packages/text-input/CHANGELOG.md
@@ -1,1 +1,16 @@
-2.0.0 - add `InputButton` component to package `text-input`.
+# Changelog
+
+## 2.1.1 - [19-02-2019]
+### Changed
+- Updated Input props to include all the other input html attributes including `autoFocus`
+- Added inputSize prop which is the value for the html input size attribute
+
+
+## 2.0.0
+### Changed
+- add `InputButton` component to package `text-input`.
+
+
+
+
+